Mejora 3: soporte para coeficientes decimales

+++ src/ecuacionsegundogradoramas/Main.java
@@ -7,17 +7,17 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int a, b, c;
+        double a, b, c;
         double discriminante, x1, x2;
 
         System.out.print("Ingrese a: ");
-        a = sc.nextInt();
+        a = sc.nextDouble();
 
         System.out.print("Ingrese b: ");
-        b = sc.nextInt();
+        b = sc.nextDouble();
 
         System.out.print("Ingrese c: ");
-        c = sc.nextInt();
+        c = sc.nextDouble();
 
         if (a == 0) {
             System.out.println("No es una ecuacion de segundo grado.");
@@ -29,7 +29,7 @@ public class Main {
         if (discriminante > 0) {
             x1 = (-b + Math.sqrt(discriminante)) / (2.0 * a);
             x2 = (-b - Math.sqrt(discriminante)) / (2.0 * a);
-
+            
             System.out.println("Dos raices reales diferentes:");
             System.out.println("Raiz 1: " + x1);
             System.out.println("Raiz 2: " + x2);
